requirements.txt
nltk
flask
numpy==1.19.2
tensorflow
pickle-mixin
python-dotenv